refactor for future happiness
