+++ Day5/Ruby/string_judge_test.rb
@@ -1,17 +1,17 @@
 require 'test/unit'
-require_relative 'string_judger.rb'
+require_relative 'string_judge.rb'
 require 'pry'
 
 # test the string judger class
 class StringJudgerTest < Test::Unit::TestCase
   def test_input
-    sj = StringJudger.new
+    # input is examples from the instructions + one edge case
     input = "qjhvhtzxzqqjkmpb\nxxyxx\nuurcxstgmygtbstg\nieodomkazucvgmuy\nmmggfwapsetemiuj"
-    assert_equal 2, sj.judge_input(input)
+    assert_equal 2, StringJudge.new.judge_input(input)
   end
 
   def test_input_legacy
     input = "aaa\nab" # shortest nice followed by shortest naughty - legacy
-    assert_equal 1, StringJudger.new.legacy_judge_input(input)
+    assert_equal 1, StringJudge.new.judge_input(input, legacy: true)
   end
 end
